+++ CHANGELOG.md
@@ -34,7 +34,8 @@ _None._
 
 ### Breaking Changes
 
-_None._
+- `logErrorImmediately` and `logErrorsImmediately` no longer have a `Result` parameter in their callback [#232]
+- `logErrorImmediately` and `logErrorsImmediately` no longer `throws` [#236]
 
 ### New Features